# openapi.yaml
# Generated by `make openapi`. Do not edit.
openapi: 3.1.0
info:
  title: Botasaurus Scrape API
  description: 'Docker-first scrape API that uses Botasaurus to fetch rendered HTML.


    - `GET /health` — liveness and detected Botasaurus version

    - `POST /scrape` — scrape a public `http`/`https` URL


    `wait_timeout_seconds` values outside `[1, 30]`

    are **clamped** into that range so scrape still runs; they are not rejected

    with 422.


    When `html` is present it is UTF-8-normalized and `headers` `content-type` is

    `text/html; charset=utf-8`.


    Localhost, private, link-local, multicast, reserved, and unspecified

    destinations are blocked (403). Schema validation failures use this API''s

    scrape error envelope, not FastAPI `detail`.'
  contact:
    name: html2rss
    url: https://github.com/html2rss/botasaurus-scrape-api/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
servers:
- url: http://localhost:4010
  description: Local Docker (make serve)
paths:
  /health:
    get:
      tags:
      - health
      summary: Health
      description: Return liveness status, service name, and the installed Botasaurus
        version.
      operationId: get-health
      responses:
        '200':
          description: Service is up.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                service: botasaurus-scrape-api
                botasaurus_version: 4.0.91
  /scrape:
    post:
      tags:
      - scrape
      summary: Scrape a URL
      description: Fetch rendered HTML for a public http(s) URL. Invalid or blocked
        targets return `ScrapeError`. `wait_timeout_seconds` is clamped, not 422.
      operationId: scrape-url
      parameters:
      - name: X-Request-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Request-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrapeRequest'
      responses:
        '200':
          description: Rendered HTML plus diagnostics. `html` is UTF-8-normalized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrapeSuccess'
              example:
                url: https://example.com
                final_url: https://example.com/
                status_code: 200
                headers:
                  content-type: text/html; charset=utf-8
                html: <!doctype html><html><body>Example Domain</body></html>
                xhr_responses: []
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 1
                  render_ms: 154
                  execution_tier: http_request
                  challenge:
                    blocked: false
                    detected: false
        '400':
          description: URL rejected by validation (scheme, host, or unresolvable target).
          content:
            application/json:
              example:
                url: https://example.com
                error: Target URL is blocked
                error_category: validation
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 0
                  render_ms: 0
              schema:
                $ref: '#/components/schemas/ScrapeError'
        '403':
          description: URL blocked by SSRF guardrails (localhost, private, or reserved
            destination).
          content:
            application/json:
              example:
                url: https://example.com
                error: Target URL is blocked
                error_category: validation
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 0
                  render_ms: 0
              schema:
                $ref: '#/components/schemas/ScrapeError'
        '422':
          description: Request schema validation failed. Body is the scrape error
            envelope, not FastAPI `detail`.
          content:
            application/json:
              example:
                url: https://example.com
                error: Target URL is blocked
                error_category: validation
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 0
                  render_ms: 0
              schema:
                $ref: '#/components/schemas/ScrapeError'
        '502':
          description: Scrape execution failure or challenge block after the final
            attempt.
          content:
            application/json:
              example:
                url: https://example.com
                error: Bot challenge detected (Just a moment...)
                error_category: challenge_block
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 3
                  strategy_used: get
                  render_ms: 1500
                  execution_tier: browser_driver
                  challenge:
                    blocked: true
                    detected: true
                    marker: Just a moment...
              schema:
                $ref: '#/components/schemas/ScrapeError'
        '504':
          description: Scrape timed out before a result was produced.
          content:
            application/json:
              example:
                url: https://example.com
                error: Scrape timed out after 45 seconds
                error_category: timeout
                diagnostics:
                  request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
                  attempts: 0
                  render_ms: 0
              schema:
                $ref: '#/components/schemas/ScrapeError'
components:
  schemas:
    ChallengeSignal:
      properties:
        blocked:
          type: boolean
          title: Blocked
          description: True when HTTP status or HTML markers indicate an anti-bot
            block.
          examples:
          - false
        detected:
          type: boolean
          title: Detected
          description: True when a known challenge interstitial or driver bot-detection
            signal matched.
          examples:
          - false
        marker:
          anyOf:
          - type: string
          - type: 'null'
          title: Marker
          description: Matched challenge marker, or null when none matched.
          examples:
          - Just a moment...
      type: object
      required:
      - blocked
      - detected
      title: ChallengeSignal
    ErrorCategory:
      type: string
      enum:
      - timeout
      - challenge_block
      - navigation_error
      - metadata_error
      - validation
      title: ErrorCategory
    ExecutionMode:
      type: string
      enum:
      - auto
      - request
      - browser
      title: ExecutionMode
    ExecutionTier:
      type: string
      enum:
      - http_request
      - browser_driver
      title: ExecutionTier
    HealthResponse:
      properties:
        status:
          type: string
          const: ok
          title: Status
          description: Liveness marker. Always `ok` when the process can serve requests.
          examples:
          - ok
        service:
          type: string
          title: Service
          description: Service identity.
          examples:
          - botasaurus-scrape-api
        botasaurus_version:
          type: string
          title: Botasaurus Version
          description: Installed Botasaurus package version, or `unknown` if metadata
            is missing.
          examples:
          - 4.0.91
      type: object
      required:
      - status
      - service
      - botasaurus_version
      title: HealthResponse
      examples:
      - botasaurus_version: 4.0.91
        service: botasaurus-scrape-api
        status: ok
    NavigationMode:
      type: string
      enum:
      - auto
      - get
      - google_get
      - google_get_bypass
      - organic_get
      title: NavigationMode
    ScrapeDiagnostics:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Unique id for this scrape attempt, used for tracing and runtime
            isolation.
          examples:
          - b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
        attempts:
          type: integer
          title: Attempts
          description: Number of navigation or request attempts actually performed.
          default: 0
          examples:
          - 1
        strategy_used:
          anyOf:
          - $ref: '#/components/schemas/NavigationMode'
          - type: 'null'
          description: Browser navigation strategy used on the final attempt. Null
            for the HTTP-request tier and for failures before navigation starts.
          examples:
          - google_get
        render_ms:
          type: integer
          title: Render Ms
          description: Elapsed scrape runtime in milliseconds.
          default: 0
          examples:
          - 154
        execution_tier:
          anyOf:
          - $ref: '#/components/schemas/ExecutionTier'
          - type: 'null'
          description: Tier that produced this result, or null when execution never
            started.
          examples:
          - http_request
        challenge:
          anyOf:
          - $ref: '#/components/schemas/ChallengeSignal'
          - type: 'null'
          description: Anti-bot assessment for this attempt, or null when detection
            did not run.
          examples:
          - blocked: false
            detected: false
        timeout_phase:
          anyOf:
          - $ref: '#/components/schemas/TimeoutPhase'
          - type: 'null'
          description: 'When `error_category` is `timeout`, which stage burned the
            budget: `queue` (threadpool wait), `boot` (browser/driver start), or `work`
            (navigate/wait/scroll). Null for non-timeout outcomes.'
          examples:
          - work
      type: object
      required:
      - request_id
      title: ScrapeDiagnostics
    ScrapeError:
      properties:
        url:
          type: string
          title: Url
          description: Requested scrape URL as submitted, or empty when the body had
            no URL.
          examples:
          - https://example.com
        error:
          type: string
          title: Error
          description: Human-readable failure message.
          examples:
          - Target URL is blocked
        error_category:
          $ref: '#/components/schemas/ErrorCategory'
          description: 'Closed failure class: `timeout`, `challenge_block`, `navigation_error`,
            `metadata_error`, or `validation`.'
          examples:
          - validation
        diagnostics:
          $ref: '#/components/schemas/ScrapeDiagnostics'
          description: Per-request tracing, strategy, timing, and challenge signals.
          examples:
          - attempts: 0
            render_ms: 0
            request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
      type: object
      required:
      - url
      - error
      - error_category
      - diagnostics
      title: ScrapeError
    ScrapeRequest:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
          description: Absolute http(s) URL to scrape. Localhost and private destinations
            are rejected by SSRF guardrails.
          examples:
          - https://example.com
        execution_mode:
          $ref: '#/components/schemas/ExecutionMode'
          description: '`auto` tries anti-detect HTTP first and escalates to the browser;
            `request` stays on HTTP; `browser` always uses Chromium.'
          default: auto
          examples:
          - auto
        navigation_mode:
          $ref: '#/components/schemas/NavigationMode'
          description: Browser navigation strategy. `auto` tries `google_get`, then
            `google_get_bypass`, then `get`.
          default: auto
          examples:
          - auto
        max_retries:
          type: integer
          maximum: 3.0
          minimum: 0.0
          title: Max Retries
          description: Retries after the first attempt (`attempts = 1 + max_retries`).
            `auto` is capped at three strategy steps.
          default: 2
          examples:
          - 2
        wait_for_selector:
          anyOf:
          - type: string
          - type: 'null'
          title: Wait For Selector
          description: CSS selector to wait for before capture. When set, execution
            uses the browser tier.
          examples:
          - h1
        wait_timeout_seconds:
          type: integer
          title: Wait Timeout Seconds
          description: Selector wait timeout in seconds. Values outside [1, 30] are
            clamped into that range so scrape still runs; they are not rejected with
            422.
          default: 15
          examples:
          - 15
        scroll:
          type: boolean
          title: Scroll
          description: Scroll to the bottom to trigger lazy-loaded content. Routes
            to the browser tier when true.
          default: false
          examples:
          - false
        block_images:
          type: boolean
          title: Block Images
          description: Ask the driver to skip image downloads.
          default: true
          examples:
          - true
        block_images_and_css:
          type: boolean
          title: Block Images And Css
          description: Ask the driver to skip image and CSS downloads.
          default: false
          examples:
          - false
        block_trackers:
          type: boolean
          title: Block Trackers
          description: Block tracking/ad networks and web fonts to speed up rendering.
          default: true
          examples:
          - true
        wait_for_complete_page_load:
          type: boolean
          title: Wait For Complete Page Load
          description: Wait for the driver complete-page-load signal before capture.
          default: true
          examples:
          - true
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
          description: Explicit User-Agent. Overrides a User-Agent header when both
            are set.
          examples:
          - Mozilla/5.0 (compatible; html2rss)
        headers:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Headers
          description: Extra HTTP headers forwarded to the request client or browser
            session.
          examples:
          - Accept-Language: en-US,en;q=0.9
        cookies:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Cookies
          description: Cookie name/value map forwarded to the request client or browser
            session.
          examples:
          - session: abc123
        window_size:
          anyOf:
          - $ref: '#/components/schemas/WindowSize'
          - type: 'null'
          description: Browser viewport size passed to the driver.
          examples:
          - height: 1080
            width: 1920
        lang:
          anyOf:
          - type: string
          - type: 'null'
          title: Lang
          description: Browser language passed to the driver.
          examples:
          - en-US
        headless:
          type: boolean
          title: Headless
          description: Run Chromium headless. Default false uses a virtual display
            in Docker.
          default: false
          examples:
          - false
        proxy:
          anyOf:
          - type: string
          - type: 'null'
          title: Proxy
          description: Proxy URL passed to the driver. Invalid or blocked proxy URLs
            are rejected by SSRF guardrails.
          examples:
          - http://user:pass@proxy.example:8080
      type: object
      required:
      - url
      title: ScrapeRequest
      examples:
      - url: https://example.com
      - block_images: true
        block_images_and_css: false
        block_trackers: true
        cookies:
          session: abc123
        execution_mode: auto
        headers:
          Accept-Language: en-US,en;q=0.9
        headless: false
        lang: en-US
        max_retries: 2
        navigation_mode: auto
        proxy: http://user:pass@proxy.example:8080
        scroll: true
        url: https://example.com
        user_agent: Mozilla/5.0 (compatible; html2rss)
        wait_for_complete_page_load: true
        wait_for_selector: h1
        wait_timeout_seconds: 15
        window_size:
          height: 1080
          width: 1920
    ScrapeSuccess:
      properties:
        url:
          type: string
          title: Url
          description: Requested scrape URL as submitted.
          examples:
          - https://example.com
        final_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Url
          description: Best-effort landing URL after redirects.
          examples:
          - https://example.com/
        status_code:
          anyOf:
          - type: integer
          - type: 'null'
          title: Status Code
          description: Best-effort HTTP status of the main document.
          examples:
          - 200
        headers:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Headers
          description: Best-effort response headers. When `html` is present, `content-type`
            is `text/html; charset=utf-8`.
          examples:
          - content-type: text/html; charset=utf-8
        html:
          type: string
          title: Html
          description: Rendered page HTML, UTF-8-normalized. When present, headers
            `content-type` is `text/html; charset=utf-8`.
          examples:
          - <!doctype html><html><body>Example Domain</body></html>
        metadata_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Metadata Error
          description: Metadata extraction failure message. HTML success is still
            returned.
          examples:
          - null
        xhr_responses:
          items:
            $ref: '#/components/schemas/XhrResponse'
          type: array
          title: Xhr Responses
          description: JSON XHR/fetch sub-resource bodies captured on the browser
            tier (empty on the HTTP-request tier).
          examples:
          - []
        diagnostics:
          $ref: '#/components/schemas/ScrapeDiagnostics'
          description: Per-request tracing, strategy, timing, and challenge signals.
          examples:
          - attempts: 1
            challenge:
              blocked: false
              detected: false
            execution_tier: http_request
            render_ms: 154
            request_id: b01ef2f8-f641-4e75-8ef2-0b73f7b4f372
      type: object
      required:
      - url
      - html
      - diagnostics
      title: ScrapeSuccess
    TimeoutPhase:
      type: string
      enum:
      - queue
      - boot
      - work
      title: TimeoutPhase
    WindowSize:
      properties:
        width:
          type: integer
          minimum: 1.0
          title: Width
          description: Browser viewport width in pixels.
          examples:
          - 1920
        height:
          type: integer
          minimum: 1.0
          title: Height
          description: Browser viewport height in pixels.
          examples:
          - 1080
      type: object
      required:
      - width
      - height
      title: WindowSize
    XhrResponse:
      properties:
        url:
          type: string
          title: Url
          description: Sub-resource URL whose JSON body was captured.
          examples:
          - https://api.example.com/items
        status_code:
          type: integer
          title: Status Code
          description: HTTP status code of the captured XHR/fetch response.
          examples:
          - 200
        headers:
          additionalProperties:
            type: string
          type: object
          title: Headers
          description: Allowlisted response headers. Only `content-type` is kept.
          examples:
          - content-type: application/json
        body:
          type: string
          title: Body
          description: Raw JSON response body as UTF-8 text.
          examples:
          - '{"items":[]}'
      type: object
      required:
      - url
      - status_code
      - body
      title: XhrResponse
tags:
- name: health
  description: Liveness probe and detected Botasaurus package version.
- name: scrape
  description: Render a public URL and return UTF-8 HTML plus diagnostics.